update cpp submodule (#24)
This also tests all of our CI to ensure the process works.

* Some good overall practices -
[guide](https://gist.github.com/nyurik/4e299ad832fd2dd43d2b27191ed3ec30)
(I put it together, so there might be some mistakes)
* It is usually best to let the author of the PR merge it themselves IF
they have the rights to do so. But only after it gets approved. This way
the author stays responsible for their work. If a contribution is from
an outside contributor, one of the maintainers must take over that work.

There are a few additional things we will need to do - i'll create
issues for those